Layer 136, allow garbage collection
