+++ src/danog/MadelineProto/API.php
@@ -19,18 +19,13 @@
 
 namespace danog\MadelineProto;
 
-use Amp\Deferred;
-use Amp\Ipc\Sync\ChannelledSocket;
 use Amp\Loop;
-use danog\MadelineProto\Ipc\Client;
-use danog\MadelineProto\Ipc\Server;
-use danog\MadelineProto\Settings\Ipc as SettingsIpc;
-use danog\MadelineProto\Settings\Logger as SettingsLogger;
+use danog\MadelineProto\Doc\MethodDoc;
 
 /**
  * Main API wrapper for MadelineProto.
  */
-class API extends InternalDoc
+final class API extends MethodDoc
 {
     /**
      * Release version.
@@ -86,48 +81,8 @@ class API extends InternalDoc
      * @var int
      */
     const SECRET_READY = MTProto::SECRET_READY;
+
     use \danog\Serializable;
-    use \danog\MadelineProto\ApiWrappers\Start;
-    use \danog\MadelineProto\ApiWrappers\Templates;
-    /**
-     * Session paths.
-     *
-     * @internal
-     * @var SessionPaths
-     */
-    public $session;
-
-    /**
-     * Instance of MadelineProto.
-     *
-     * @var null|MTProto|Client
-     */
-    public $API;
-
-    /**
-     * Storage for externally set properties to be serialized.
-     *
-     * @var array
-     */
-    protected array $storage = [];
-
-    /**
-     * Whether we're getting our API ID.
-     *
-     * @internal
-     *
-     * @var boolean
-     */
-    private bool $gettingApiId = false;
-
-    /**
-     * my.telegram.org API wrapper.
-     *
-     * @internal
-     *
-     * @var null|MyTelegramOrgWrapper
-     */
-    private $myTelegramOrgWrapper;
 
     /**
      * Whether this is an old instance.
@@ -141,13 +96,6 @@ class API extends InternalDoc
      * @var boolean
      */
     private bool $destructing = false;
-
-    /**
-     * API wrapper (to avoid circular references).
-     *
-     * @var APIWrapper
-     */
-    private $wrapper;
 
     /**
      * Unlock callback.
@@ -167,190 +115,16 @@ class API extends InternalDoc
     public function __magic_construct(string $session, $settings = []): void
     {
         Magic::start(true);
-        $settings = Settings::parseFromLegacy($settings);
-        $this->session = new SessionPaths($session);
-        $this->wrapper = new APIWrapper($this, $this->exportNamespace());
-
-        $this->setInitPromise($this->internalInitAPI($settings));
-        foreach (\get_class_vars(APIFactory::class) as $key => $var) {
-            if (\in_array($key, ['namespace', 'API', 'lua', 'async', 'asyncAPIPromise', 'methods'])) {
-                continue;
-            }
-            if (!$this->{$key}) {
-                $this->{$key} = $this->exportNamespace($key);
-            }
-        }
+        $this->initProxyNamespaces(new APIWrapper(new SessionPaths($session), Settings::parseFromLegacy($settings)));
     }
     /**
-     * Async constructor function.
+     * Sleep function.
      *
-     * @param Settings|SettingsEmpty|SettingsIpc $settings Settings
-     *
-     * @return \Generator
+     * @return array
      */
-    private function internalInitAPI(SettingsAbstract $settings): \Generator
+    public function __sleep()
     {
-        Logger::constructorFromSettings($settings instanceof Settings
-            ? $settings->getLogger()
-            : ($settings instanceof SettingsLogger ? $settings : new SettingsLogger));
-
-        if (yield from $this->connectToMadelineProto($settings)) {
-            return; // OK
-        }
-
-        if (!$settings instanceof Settings) {
-            $newSettings = new Settings;
-            $newSettings->merge($settings);
-            $settings = $newSettings;
-        }
-
-        $appInfo = $settings->getAppInfo();
-        if (!$appInfo->hasApiInfo()) {
-            $app = yield from $this->APIStart($settings);
-            if (!$app) {
-                $this->forceInit(true);
-                die();
-            }
-            $appInfo->setApiId($app['api_id']);
-            $appInfo->setApiHash($app['api_hash']);
-        }
-        $this->API = new MTProto($settings, $this->wrapper);
-        yield from $this->API->initAsynchronously();
-        $this->APIFactory();
-        $this->logger->logger(Lang::$current_lang['madelineproto_ready'], Logger::NOTICE);
-    }
-
-    /**
-     * Reconnect to full instance.
-     *
-     * @return \Generator
-     */
-    protected function reconnectFull(): \Generator
-    {
-        if (!$this->API) {
-            yield from $this->initAsynchronously();
-        }
-        if ($this->API instanceof Client) {
-            $this->logger->logger("Restarting to full instance...");
-            try {
-                if (!isset($_GET['MadelineSelfRestart']) && ((yield $this->hasEventHandler()) || !(yield $this->isIpcWorker()))) {
-                    $this->logger->logger("Restarting to full instance: the bot is already running!");
-                    Tools::closeConnection(yield $this->getWebMessage("The bot is already running!"));
-                    return false;
-                }
-                $this->logger->logger("Restarting to full instance: stopping IPC server...");
-                yield $this->API->stopIpcServer();
-                $this->logger->logger("Restarting to full instance: disconnecting from IPC server...");
-                yield $this->API->disconnect();
-            } catch (SecurityException $e) {
-                throw $e;
-            } catch (\Throwable $e) {
-                $this->logger->logger("Restarting to full instance: error $e");
-            }
-            $this->logger->logger("Restarting to full instance: reconnecting...");
-            $cancel = new Deferred;
-            $cb = function () use ($cancel, &$cb): \Generator {
-                [$result] = yield from Serialization::tryConnect($this->session->getIpcPath(), $cancel->promise());
-                if ($result instanceof ChannelledSocket) {
-                    try {
-                        if (!$this->API instanceof Client) {
-                            $this->logger->logger("Restarting to full instance (again): the bot is already running!");
-                            yield $result->disconnect();
-                            return;
-                        }
-                        $API = new Client($result, $this->session, Logger::$default, $this->async);
-                        if ((yield from $API->hasEventHandler()) || !(yield from $API->isIpcWorker())) {
-                            $this->logger->logger("Restarting to full instance (again): the bot is already running!");
-                            yield $API->disconnect();
-                            $API->unreference();
-                            return;
-                        }
-                        $this->logger->logger("Restarting to full instance: stopping another IPC server...");
-                        yield $API->stopIpcServer();
-                        $this->logger->logger("Restarting to full instance: disconnecting from IPC server...");
-                        yield $API->disconnect();
-                        $API->unreference();
-                    } catch (SecurityException $e) {
-                        throw $e;
-                    } catch (\Throwable $e) {
-                        $this->logger->logger("Restarting to full instance: error in stop loop $e");
-                    }
-                    Tools::callFork($cb());
-                }
-            };
-            Tools::callFork($cb());
-            yield from $this->connectToMadelineProto(new SettingsEmpty, true);
-            $cancel->resolve(new Exception('Connected!'));
-        }
-        return true;
-    }
-    /**
-     * Connect to MadelineProto.
-     *
-     * @param SettingsAbstract $settings Settings
-     * @param bool $forceFull Whether to force full initialization
-     *
-     * @return \Generator
-     */
-    protected function connectToMadelineProto(SettingsAbstract $settings, bool $forceFull = false, bool $tryReconnect = true): \Generator
-    {
-        if ($settings instanceof SettingsIpc) {
-            $forceFull = $forceFull || $settings->getSlow();
-        } elseif ($settings instanceof Settings) {
-            $forceFull = $forceFull || $settings->getIpc()->getSlow();
-        }
-        $forceFull = $forceFull || isset($_GET['MadelineSelfRestart']) || Magic::$altervista;
-
-        [$unserialized, $this->unlock] = yield Tools::timeoutWithDefault(
-            Serialization::unserialize($this->session, $settings, $forceFull),
-            30000,
-            [0, null]
-        );
-
-        if ($unserialized === 0) {
-            // Timeout
-            Logger::log("!!! Could not connect to MadelineProto, please check and report the logs for more details. !!!", Logger::FATAL_ERROR);
-            if (!$tryReconnect || (\defined('MADELINEPROTO_TEST') && \constant("MADELINEPROTO_TEST") === 'testing')) {
-                throw new Exception('Could not connect to MadelineProto, please check the MadelineProto.log file to debug!');
-            }
-            Logger::log("!!! Reconnecting using slower method. !!!", Logger::FATAL_ERROR);
-            // IPC server error, try fetching full session
-            return yield from $this->connectToMadelineProto($settings, true, false);
-        } elseif ($unserialized instanceof \Throwable) {
-            // IPC server error, try fetching full session
-            return yield from $this->connectToMadelineProto($settings, true);
-        } elseif ($unserialized instanceof ChannelledSocket) {
-            // Success, IPC client
-            $this->API = new Client($unserialized, $this->session, Logger::$default, $this->async);
-            $this->APIFactory();
-            return true;
-        } elseif ($unserialized) {
-            // Success, full session
-            if ($this->API) {
-                $this->API->unreference();
-                $this->API = null;
-            }
-            $unserialized->storage = $unserialized->storage ?? [];
-            $unserialized->session = $this->session;
-            APIWrapper::link($this, $unserialized);
-            APIWrapper::link($this->wrapper, $this);
-            AbstractAPIFactory::link($this->wrapper->getFactory(), $this);
-            if (isset($this->API)) {
-                $this->storage = $this->API->storage ?? $this->storage;
-
-                unset($unserialized);
-
-                if ($settings instanceof SettingsIpc) {
-                    $settings = new SettingsEmpty;
-                }
-                $this->methods = self::getInternalMethodList($this->API, MTProto::class);
-                yield from $this->API->wakeup($settings, $this->wrapper);
-                $this->APIFactory();
-                $this->logger->logger(Lang::$current_lang['madelineproto_ready'], Logger::NOTICE);
-                return true;
-            }
-        }
-        return false;
+        return [];
     }
     /**
      * Wakeup function.
@@ -368,21 +142,11 @@ class API extends InternalDoc
      */
     public function __destruct()
     {
-        $this->init();
+        $this->wrapper->init();
         if (!$this->oldInstance) {
             $this->logger->logger('Shutting down MadelineProto ('.static::class.')');
             $this->destructing = true;
-            if ($this->API) {
-                if ($this->API instanceof Tools) {
-                    $this->API->destructing = true;
-                }
-                $this->API->unreference();
-            }
-            if (isset($this->wrapper) && (!Magic::$signaled || $this->gettingApiId)) {
-                $this->logger->logger('Prompting final serialization...');
-                Tools::wait($this->wrapper->serialize());
-                $this->logger->logger('Done final serialization!');
-            }
+            $this->wrapper->destruct();
             if ($this->unlock) {
                 ($this->unlock)();
             }
@@ -390,25 +154,6 @@ class API extends InternalDoc
             $this->logger->logger('Shutting down MadelineProto (old deserialized instance of API)');
         }
     }
-    /**
-     * Init API wrapper.
-     *
-     * @return void
-     */
-    private function APIFactory(): void
-    {
-        if ($this->API && $this->API->inited()) {
-            if ($this->API instanceof MTProto) {
-                foreach ($this->API->getMethodNamespaces() as $namespace) {
-                    if (!$this->{$namespace}) {
-                        $this->{$namespace} = $this->exportNamespace($namespace);
-                    }
-                }
-            }
-            $this->methods = self::getInternalMethodList($this->API, MTProto::class);
-        }
-    }
-
 
     /**
      * Start MadelineProto and the event handler (enables async).
@@ -425,7 +170,7 @@ class API extends InternalDoc
         $started = false;
         while (true) {
             try {
-                Tools::wait($this->startAndLoopAsyncInternal($eventHandler, $started));
+                Tools::wait($this->wrapper->startAndLoopAsyncInternal($eventHandler, $started));
                 return;
             } catch (SecurityException $e) {
                 throw $e;
@@ -433,7 +178,7 @@ class API extends InternalDoc
                 $t = \time();
                 $errors = [$t => $errors[$t] ?? 0];
                 $errors[$t]++;
-                if ($errors[$t] > 10 && (!$this->inited() || !$started)) {
+                if ($errors[$t] > 10 && (!$this->wrapper->inited() || !$started)) {
                     $this->logger->logger("More than 10 errors in a second and not inited, exiting!", Logger::FATAL_ERROR);
                     return;
                 }
@@ -465,7 +210,7 @@ class API extends InternalDoc
                 $promises = [];
                 foreach ($instances as $k => $instance) {
                     $instance->start(['async' => false]);
-                    $promises []= $instance->startAndLoopAsyncInternal($eventHandler[$k], $started[$k]);
+                    $promises []= $instance->wrapper->startAndLoopAsyncInternal($eventHandler[$k], $started[$k]);
                 }
                 Tools::wait(Tools::all($promises));
                 return;
@@ -497,50 +242,9 @@ class API extends InternalDoc
     public function startAndLoopAsync(string $eventHandler): \Generator
     {
         $started = false;
-        return $this->startAndLoopAsyncInternal($eventHandler, $started);
+        return $this->wrapper->startAndLoopAsyncInternal($eventHandler, $started);
     }
 
-    /**
-     * Start MadelineProto and the event handler (enables async).
-     *
-     * Also initializes error reporting, catching and reporting all errors surfacing from the event loop.
-     *
-     * @param string $eventHandler Event handler class name
-     *
-     * @return \Generator
-     */
-    private function startAndLoopAsyncInternal(string $eventHandler, bool &$started): \Generator
-    {
-        $this->async(true);
-
-        yield $this->start();
-        if (!yield from $this->reconnectFull()) {
-            return;
-        }
-
-        $errors = [];
-        while (true) {
-            try {
-                yield $this->setEventHandler($eventHandler);
-                $started = true;
-                /** @var API $this->API */
-                return yield from $this->API->loop();
-            } catch (SecurityException $e) {
-                throw $e;
-            } catch (\Throwable $e) {
-                $t = \time();
-                $errors = [$t => $errors[$t] ?? 0];
-                $errors[$t]++;
-                if ($errors[$t] > 10 && (!$this->inited() || !$started)) {
-                    $this->logger->logger("More than 10 errors in a second and not inited, exiting!", Logger::FATAL_ERROR);
-                    return;
-                }
-                echo $e;
-                $this->logger->logger((string) $e, Logger::FATAL_ERROR);
-                $this->report("Surfaced: $e");
-            }
-        }
-    }
     /**
      * Get attribute.
      *
@@ -552,13 +256,7 @@ class API extends InternalDoc
      */
     public function &__get(string $name)
     {
-        if ($name === 'logger') {
-            if (isset($this->API)) {
-                return $this->API->logger;
-            }
-            return Logger::$default;
-        }
-        return $this->storage[$name];
+        return $this->wrapper->getStorage($name);
     }
     /**
      * Set an attribute.
@@ -572,7 +270,7 @@ class API extends InternalDoc
      */
     public function __set(string $name, $value)
     {
-        return $this->storage[$name] = $value;
+        return $this->wrapper->setStorage($name, $value);
     }
     /**
      * Whether an attribute exists.
@@ -583,7 +281,7 @@ class API extends InternalDoc
      */
     public function __isset(string $name): bool
     {
-        return isset($this->storage[$name]);
+        return $this->wrapper->issetStorage($name);
     }
     /**
      * Unset attribute.
@@ -594,6 +292,16 @@ class API extends InternalDoc
      */
     public function __unset(string $name): void
     {
-        unset($this->storage[$name]);
+        $this->wrapper->unsetStorage($name);
+    }
+
+    /**
+     * Info to dump.
+     *
+     * @return array
+     */
+    public function __debugInfo(): array
+    {
+        return $this->wrapper->__debugInfo();
     }
 }
