--- src/danog/MadelineProto/APIFactory.php
@@ -1,126 +1,0 @@
-<?php
-
-/**
- * APIFactory module.
- *
- * This file is part of MadelineProto.
- * MadelineProto is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
- * MadelineProto is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Affero General Public License for more details.
- * You should have received a copy of the GNU General Public License along with MadelineProto.
- * If not, see <http://www.gnu.org/licenses/>.
- *
- * @author    Daniil Gentili <daniil@daniil.it>
- * @copyright 2016-2020 Daniil Gentili <daniil@daniil.it>
- * @license   https://opensource.org/licenses/AGPL-3.0 AGPLv3
- *
- * @link https://docs.madelineproto.xyz MadelineProto documentation
- */
-
-namespace danog\MadelineProto;
-
-class APIFactory extends AbstractAPIFactory
-{
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var stats
-     */
-    public $stats;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var folders
-     */
-    public $folders;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var langpack
-     */
-    public $langpack;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var phone
-     */
-    public $phone;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var stickers
-     */
-    public $stickers;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var payments
-     */
-    public $payments;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var bots
-     */
-    public $bots;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var channels
-     */
-    public $channels;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var help
-     */
-    public $help;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var upload
-     */
-    public $upload;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var photos
-     */
-    public $photos;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var updates
-     */
-    public $updates;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var messages
-     */
-    public $messages;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var contacts
-     */
-    public $contacts;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var users
-     */
-    public $users;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var account
-     */
-    public $account;
-    /**
-     * @internal this is a internal property generated by build_docs.php, don't change manually
-     *
-     * @var auth
-     */
-    public $auth;
-}
