+++ src/danog/MadelineProto/AnnotationsBuilder.php
@@ -22,40 +22,29 @@ namespace danog\MadelineProto;
 use Amp\Promise;
 use danog\MadelineProto\Settings\TLSchema;
 use danog\MadelineProto\TL\TL;
-use danog\MadelineProto\TL\TLCallback;
-use phpDocumentor\Reflection\DocBlockFactory;
 
 class AnnotationsBuilder
 {
-    /**
-     * Reflection classes.
-     */
-    private array $reflectionClasses = [];
-    /**
-     * Logger.
-     */
-    private Logger $logger;
-    /**
-     * Namespace.
-     */
-    private string $namespace;
     /**
      * TL instance.
      */
     private TL $TL;
     /**
-     * Settings.
+     * Constructor.
+     *
+     * @param Logger $logger Logger
+     * @param array $settings Settings
+     * @param string $basedir Output directory
+     * @param class-string $API API class
+     * @param class-string $MTProto MTProto class
+     * @param string $namespace Output namespace
      */
-    private array $settings;
-    /**
-     * Output file.
-     */
-    private string $output;
-    public function __construct(Logger $logger, array $settings, string $output, array $reflectionClasses, string $namespace)
+    public function __construct(Logger $logger, array $settings, private string $basedir, private string $API, private string $MTProto, private string $namespace)
     {
-        $this->reflectionClasses = $reflectionClasses;
-        $this->logger = $logger;
-        $this->namespace = $namespace;
+        $this->basedir .= "/".\str_replace('\\', '/', $namespace).'/';
+        if (!\file_exists($this->basedir)) {
+            \mkdir($this->basedir);
+        }
         /** @psalm-suppress InvalidArgument */
         $this->TL = new TL(new class($logger) {
             public Logger $logger;
@@ -67,40 +56,11 @@ class AnnotationsBuilder
         $tlSchema = new TLSchema;
         $tlSchema->mergeArray($settings);
         $this->TL->init($tlSchema);
-        $this->settings = $settings;
-        $this->output = $output;
     }
-    public function mkAnnotations(): void
+    public function run(): void
     {
         \danog\MadelineProto\Logger::log('Generating annotations...', \danog\MadelineProto\Logger::NOTICE);
-        $this->setProperties();
         $this->createInternalClasses();
-    }
-    /**
-     * Open file of class APIFactory
-     * Insert properties
-     * save the file with new content.
-     *
-     * @return void
-     */
-    private function setProperties()
-    {
-        \danog\MadelineProto\Logger::log('Generating properties...', \danog\MadelineProto\Logger::NOTICE);
-        $fixture = DocBlockFactory::createInstance();
-        $class = new \ReflectionClass($this->reflectionClasses['APIFactory']);
-        $content = \file_get_contents($filename = $class->getFileName());
-        foreach ($class->getProperties() as $property) {
-            if ($raw_docblock = $property->getDocComment()) {
-                $docblock = $fixture->create($raw_docblock);
-                if ($docblock->hasTag('internal')) {
-                    $content = \str_replace("\n    ".$raw_docblock."\n    public \$".$property->getName().';', '', $content);
-                }
-            }
-        }
-        foreach ($this->TL->getMethodNamespaces() as $namespace) {
-            $content = \preg_replace('/(class( \\w+[,]?){0,}\\n{\\n)/', '${1}'."    /**\n"."     * @internal this is a internal property generated by build_docs.php, don't change manually\n"."     *\n"."     * @var {$namespace}\n"."     */\n"."    public \${$namespace};\n", $content);
-        }
-        \file_put_contents($filename, $content);
     }
     /**
      * Create internalDoc.
@@ -110,21 +70,6 @@ class AnnotationsBuilder
     private function createInternalClasses(): void
     {
         \danog\MadelineProto\Logger::log('Creating internal classes...', \danog\MadelineProto\Logger::NOTICE);
-        $handle = \fopen($this->output, 'w');
-        \fwrite($handle, "<?php namespace {$this->namespace}; class InternalDoc extends APIFactory {}");
-        $class = new \ReflectionClass($this->reflectionClasses['API']);
-        $methods = $class->getMethods(\ReflectionMethod::IS_STATIC | \ReflectionMethod::IS_PUBLIC);
-        $ignoreMethods = ['fetchserializableobject'];
-        foreach ($methods as $method) {
-            $ignoreMethods[$method->getName()] = $method->getName();
-        }
-        $class = new \ReflectionClass(TLCallback::class);
-        $methods = $class->getMethods(\ReflectionMethod::IS_STATIC | \ReflectionMethod::IS_PUBLIC);
-        foreach ($methods as $method) {
-            $ignoreMethods[$method->getName()] = $method->getName();
-        }
-        \fclose($handle);
-        $handle = \fopen($this->output, 'w');
         $internalDoc = [];
         foreach ($this->TL->getMethods()->by_id as $id => $data) {
             if (!\strpos($data['method'], '.')) {
@@ -171,7 +116,8 @@ class AnnotationsBuilder
             }
             $internalDoc[$namespace][$method]['return'] = $type;
         }
-        $class = new \ReflectionClass($this->reflectionClasses['MTProto']);
+        $namespaces = \array_keys($internalDoc);
+        $class = new \ReflectionClass($this->MTProto);
         $methods = $class->getMethods((\ReflectionMethod::IS_STATIC & \ReflectionMethod::IS_PUBLIC) | \ReflectionMethod::IS_PUBLIC);
         $class = new \ReflectionClass(Tools::class);
         $methods = \array_merge($methods, $class->getMethods((\ReflectionMethod::IS_STATIC & \ReflectionMethod::IS_PUBLIC) | \ReflectionMethod::IS_PUBLIC));
@@ -199,7 +145,7 @@ class AnnotationsBuilder
 
         foreach ($methods as $method) {
             $name = $method->getName();
-            if (isset($ignoreMethods[$name])) {
+            if ($name === 'fetchserializableobject' || $name === 'initDb' || $name === 'deinitDb') {
                 continue;
             }
             if (\strpos($method->getDocComment() ?? '', '@internal') !== false) {
@@ -285,7 +231,7 @@ class AnnotationsBuilder
                 if (!$type->isBuiltin()) {
                     $doc .= '\\';
                 }
-                $doc .= $type->getName() === 'self' ? $this->reflectionClasses['API'] : $type->getName();
+                $doc .= $type->getName() === 'self' ? $this->API : $type->getName();
                 $async = false;
             }
             if ($method->getDeclaringClass()->getName() == Tools::class) {
@@ -296,8 +242,9 @@ class AnnotationsBuilder
             $doc .= "\n{\n";
             if ($async) {
                 $doc .= "    {$ret} \$this->__call(__FUNCTION__, {$finalParamList});\n";
-            } elseif (!$static) {
+            /*} elseif (!$static) {
                 $doc .= "    {$ret} \$this->API->{$name}({$paramList});\n";
+            */
             } else {
                 $doc .= "    {$ret} \\".$method->getDeclaringClass()->getName()."::".$name."({$paramList});\n";
             }
@@ -327,7 +274,7 @@ class AnnotationsBuilder
                     if (!$type->isBuiltin()) {
                         $new .= '\\';
                     }
-                    $new .= $type->getName() === 'self' ? $this->reflectionClasses['API'] : $type->getName();
+                    $new .= $type->getName() === 'self' ? $this->API : $type->getName();
                 }
                 $phpdoc = \str_replace("@return ".$ret, "@return mixed", $phpdoc);
                 if (!\str_contains($phpdoc, '@psalm-return')) {
@@ -339,20 +286,35 @@ class AnnotationsBuilder
                 "@psalm-return $promise<$1>",
                 $phpdoc
             );
-            $internalDoc['InternalDoc'][$name]['method'] = $phpdoc;
-            $internalDoc['InternalDoc'][$name]['method'] .= "\n    ".\implode("\n    ", \explode("\n", $doc));
+            $internalDoc['MethodDoc'][$name]['method'] = $phpdoc;
+            $internalDoc['MethodDoc'][$name]['method'] .= "\n    ".\implode("\n    ", \explode("\n", $doc));
         }
-        \fwrite($handle, "<?php\n");
-        \fwrite($handle, "/**\n");
-        \fwrite($handle, " * This file is automatic generated by build_docs.php file\n");
-        \fwrite($handle, " * and is used only for autocomplete in multiple IDE\n");
-        \fwrite($handle, " * don't modify manually.\n");
-        \fwrite($handle, " */\n\n");
-        \fwrite($handle, "namespace {$this->namespace};\n");
+        $handleNamespace = \fopen($this->basedir."/NamespaceDoc.php", 'w');
+        $handleMethod = \fopen($this->basedir."/MethodDoc.php", 'w');
+        foreach ([$handleMethod, $handleNamespace] as $handle) {
+            \fwrite($handle, "<?php\n");
+            \fwrite($handle, "/**\n");
+            \fwrite($handle, " * This file is automatic generated by build_docs.php file\n");
+            \fwrite($handle, " * and is used only for autocomplete in multiple IDE\n");
+            \fwrite($handle, " * don't modify manually.\n");
+            \fwrite($handle, " */\n\n");
+            \fwrite($handle, "namespace {$this->namespace};\n");
+        }
+        \fwrite($handleNamespace, "abstract class NamespaceDoc extends \\danog\\MadelineProto\\APIProxy {\n");
+        foreach ($this->TL->getMethodNamespaces() as $namespace) {
+            \fwrite($handleNamespace, "/** @var $namespace */\npublic \$$namespace;\n");
+        }
+        \fwrite($handleNamespace, "}\n");
+
+        $namespaces = \var_export($namespaces, true);
         foreach ($internalDoc as $namespace => $methods) {
-            if ($namespace === 'InternalDoc') {
-                \fwrite($handle, "\nclass {$namespace} extends APIFactory\n{\n");
+            if ($namespace === 'MethodDoc') {
+                $handle = $handleMethod;
+                \fwrite($handle, "\nclass {$namespace} extends NamespaceDoc\n{\n");
+                \fwrite($handle, "protected function cloneProxy(): self { return new self; }\n");
+                \fwrite($handle, "protected function getAPINamespaces(): array { return $namespaces; }\n");
             } else {
+                $handle = $handleNamespace;
                 \fwrite($handle, "\ninterface {$namespace}\n{");
             }
             foreach ($methods as $method => $properties) {
