+++ src/danog/MadelineProto/MTProtoTools/ReferenceDatabase.php
@@ -19,7 +19,6 @@
 
 namespace danog\MadelineProto\MTProtoTools;
 
-use Amp\Loop;
 use danog\MadelineProto\Db\DbArray;
 use danog\MadelineProto\Db\DbPropertiesTrait;
 use danog\MadelineProto\Exception;
@@ -100,11 +99,6 @@ class ReferenceDatabase implements TLCallback
     public function init(): \Generator
     {
         yield from $this->initDb($this->API);
-
-        //Clear table on each start to fix fatals with invalid references
-        Loop::defer(fn () =>yield $this->db->clear());
-        //Clear table every day
-        Loop::repeat(24*1000*3600, fn () =>yield $this->db->clear());
     }
     public function getMethodCallbacks(): array
     {
