+++ src/danog/MadelineProto/Settings/TLSchema.php
@@ -12,11 +12,11 @@ class TLSchema extends SettingsAbstract
     /**
      * TL layer version.
      */
-    protected int $layer = 135;
+    protected int $layer = 136;
     /**
      * API schema path.
      */
-    protected string $APISchema = __DIR__.'/../TL_telegram_v135.tl';
+    protected string $APISchema = __DIR__.'/../TL_telegram_v136.tl';
     /**
      * MTProto schema path.
      */
