+++ src/danog/MadelineProto/TL/Types/Button.php
@@ -77,7 +77,7 @@ class Button implements \JsonSerializable, \ArrayAccess
         $this->button = $button;
         $this->id = $message['id'];
         $this->API = $API;
-        $this->session = $API->getWrapper()->getSession()->getLegacySessionPath();
+        $this->session = $API->getWrapper()->session->getLegacySessionPath();
     }
     /**
      * Sleep function.
